clean up test helper WIP
